fix: 컴포넌트 분리 및 안쓰는 변수 삭제
